refactor: split renderer tests
